+++ CdIts.Caffoa.Cli/Config/ModelConfig.cs
@@ -1,3 +1,4 @@
+using System.Text.RegularExpressions;
 using CdIts.Caffoa.Cli.Errors;
 
 namespace CdIts.Caffoa.Cli.Config;
@@ -9,8 +10,9 @@ public class ModelConfig
 
     public string Namespace
     {
-        get => _ns ?? throw new ConfigurationMissingException("Missing 'namespace' of configuration 'model'");
-        set => _ns = value;
+        get => _ns ?? Regex.Replace(TargetFolder.Trim('/', '.', ' '), @"\W+", ".");
+
+    set => _ns = value;
     }
 
     public string TargetFolder